Added intervals and file checks before downloading status or data files

+++ lib/vatsim/status.rb
@@ -3,16 +3,19 @@ require 'uri'
 
 module Vatsim
   class Status
+
+    @@status_download_interval = 60*60*6 # 6 hours
+    @@vatsim_data_download_interval = 60*5 # 5 minutes
     
     STATUS_FILE_PATH = File.expand_path("../../../cache", __FILE__) + "/status.txt"
     VATSIMDATA_FILE_PATH = File.expand_path("../../../cache", __FILE__) + "/vatsim-data.txt"
         
     def self.download
 
-      if !File.exists?(STATUS_FILE_PATH) or File.mtime(STATUS_FILE_PATH) < Time.now - 600
-        File.new(STATUS_FILE_PATH)
-        Net::HTTP.start("status.vatsim.net") { |http|
-          resp = http.get("/status.txt")
+      if !File.exists?(STATUS_FILE_PATH) or File.mtime(STATUS_FILE_PATH) < Time.now - @@status_download_interval 
+        File.new(STATUS_FILE_PATH, File::CREAT)
+	Net::HTTP.start("status.vatsim.net") { |http|
+       	  resp = http.get("/status.txt")
           open(STATUS_FILE_PATH, "wb") { |file|
             file.write(resp.body)
           }
@@ -29,15 +32,15 @@ module Vatsim
         end
       }
       url = URI.parse(URI.encode(url0s[rand(url0s.length)].strip!))
-      
-    File.new(VATSIMDATA_FILE_PATH)
+if !File.exists?(VATSIMDATA_FILE_PATH) or File.mtime(VATSIMDATA_FILE_PATH) < Time.now - @@vatsim_data_download_interval      
+    File.new(VATSIMDATA_FILE_PATH, File::CREAT)
     Net::HTTP.start(url.host) { |http|
       resp = http.get(url.path)
       open(VATSIMDATA_FILE_PATH, "wb") { |file|
         file.write(resp.body)
       }
     }
-      
+     end 
     end
   end
 end
